Refactored some of the code and all i need to do is to add testing

+++ src/main.rs
@@ -1,8 +1,8 @@
 use std::env;
 use std::fs;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
-fn move_to_trash(trash_path: &PathBuf, file: PathBuf) {
+fn move_to_trash(trash_path: &Path, file: PathBuf) {
     match file.file_name() {
         Some(file_name) => {
             let new_path = trash_path.join(file_name);
@@ -17,6 +17,27 @@ fn move_to_trash(trash_path: &PathBuf, file: PathBuf) {
     }
 }
 
+// need to make this into an Option type so it does not crash
+fn get_trash_path() -> PathBuf {
+    // Getting the location of the Trash folder on linux
+    let home_directoy = env::home_dir().expect("Cannot find the home direcory for this os");
+    home_directoy.join(".local/share/Trash/files/")
+}
+
+fn process_content_for_deletion(trash_path: &Path, content: &[String]) {
+    // purely exists to make testing easier and breaking up logic from main
+
+    let absolute_content = content.iter().map(fs::canonicalize); // Getting the absolute path for all the files or folders
+
+    // looping through all the files and moving them to Trash one by one
+    for item in absolute_content {
+        match item {
+            Ok(path) => move_to_trash(trash_path, path),
+            Err(e) => eprintln!("Program exited with error: {:}", e),
+        }
+    }
+}
+
 fn main() {
     let args: Vec<String> = env::args().collect();
 
@@ -25,61 +46,21 @@ fn main() {
         return;
     }
 
-    // Getting the location of the Trash folder on linux
-    let home_directoy = env::home_dir().expect("Cannot find the home direcory for this os");
-    let trash_location = home_directoy.join(".local/share/Trash/files/");
+    let trash_path = get_trash_path();
+    let trash_path = trash_path.as_path();
 
-    match args[1].as_str() {
-        "-h" => println!(
+    if args[1].as_str() == "-h" {
+        println!(
             "Safe-rm is a safe alternative to rm command. It outputs all removed files to the Trash folder."
-        ),
+        );
+    } else {
+        let mid = if args[1].as_str() == "-r" { 2 } else { 1 }; // determines when to cut the args. 1 for files and 2 for folders
+        let (_, content) = args.split_at(mid);
 
-        "-r" => {
-            // meeant for folders
-            let (_, folder_list) = args.split_at(2);
-
-            if folder_list.is_empty() {
-                println!("Please add folders to delete");
-                return;
-            }
-            //Getting the absolute path for all the files
-            let all_files = folder_list.iter().map(fs::canonicalize);
-
-            // looping through all the files and moving them to Trash one by one
-            for file_path in all_files {
-                match file_path {
-                    Ok(path) => {
-                        // This might not be nessary since move to trah moved the whole project including the subfolders
-                        // Might just need to expirement with it
-
-                        move_to_trash(&trash_location, path)
-                    }
-                    Err(e) => eprintln!("Program exited with error: {:}", e),
-                }
-            }
+        if content.is_empty() {
+            println!("No files or folders were input for deletetion");
+            return;
         }
-        _ => {
-            // Already checked if the len of the args are bigger than one
-            let (_, files) = args.split_at(1);
-
-            //Getting the absolute path for all the files
-            let all_files = files.iter().map(fs::canonicalize);
-
-            // looping through all the files and moving them to Trash one by one
-            for file_path in all_files {
-                match file_path {
-                    Ok(path) => {
-                        // This might not be nessary since move to trah moved the whole project including the subfolders
-                        // Might just need to expirement with it
-                        if path.is_file() {
-                            move_to_trash(&trash_location, path)
-                        } else {
-                            println!("{:?} is a direcory, please use the -r command", path)
-                        }
-                    }
-                    Err(e) => eprintln!("Program exited with error: {:}", e),
-                }
-            }
-        }
+        process_content_for_deletion(trash_path, content);
     }
 }
